addressed code review comments
